Шейкерная сортировка
