+++ template/src/sortdemo.c
@@ -1,3 +1,4 @@
+#include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -11,15 +12,18 @@
     \param[in] size размер массива
 */
 int* bubble_sort(int* array, int size) {
+	int swap_count = 0;
 	for (int i = 0; i < size; i++) {
 		for (int j = 0; j < size; j++) {
 			if (array[j] > array[j + 1]) {
 				int t = array[j];
 				array[j] = array[j + 1];
 				array[j + 1] = t;
+				swap_count++;
 			}
 		}
 	}
+	printf("Количество перестановок: %d\n", swap_count);
 	return array;
 }
 
@@ -56,6 +60,48 @@ int* quick_sort(int* src) {
     return 0;
 }
 
+void swap(int* array, int i, int j) {
+	int t = array[j];
+	array[j] = array[j + 1];
+	array[j + 1] = t;
+}
+
+/**
+    Сортировка слиянием.
+
+    \param[out] dest отсортированный массив
+    \param[in] array исходный массив
+    \param[in] size размер массива
+*/
+int* shaker_sort(int* array, int size) {
+	int sort_or_not = 1;
+	int right = size, left = 1;
+	int swap_count = 0;
+	do {
+		sort_or_not = 1;
+		for (int i = left; i <= right; i++) {
+			if (array[i - 1] > array[i]) {
+				swap(array, i - 1, i);
+				swap_count++;
+				sort_or_not = 0;
+			}
+		}
+		right--;
+		for (int i = right; i >= left; i--) {
+			if (array[i] < array[i - 1]) {
+				swap(array, i, i - 1);
+				swap_count++;
+				sort_or_not = false;
+			}
+		}
+		left++;
+	} while (sort_or_not == false);
+
+	printf("Количество перестановок: %d\n", swap_count);
+
+	return array;
+}
+
 int* random_array(int* array, int size, int num) {
 	for (int i = 0; i < size; i++) {
 		array[i] = rand() % num;
@@ -80,7 +126,7 @@ int main(void) {
 	srand(time(NULL));
 
 	int array_size = 50;
-	int src[array_size] = { };
+	int src[array_size];
 	int random_radix = 100;
 
 	random_array(src, array_size, random_radix);
@@ -88,10 +134,21 @@ int main(void) {
 	printf("Исходный массив: ");
 	print_array(src, array_size);
 
-	bubble_sort(src, array_size);
+	printf("Сортировка пузырьком: \n");
+	int src1[array_size];
+	memcpy(src1, src, array_size * sizeof(int));
+	bubble_sort(src1, array_size);
 
 	printf("Отсортированный массив: ");
-	print_array(src, array_size);
+	print_array(src1, array_size);
+
+	printf("Шейкерная сортировка: \n");
+	int src2[array_size];
+	memcpy(src2, src, array_size * sizeof(int));
+	shaker_sort(src2, array_size);
+
+	printf("Отсортированный массив: ");
+	print_array(src2, array_size);
 
 	printf("Статистика:\n");
 	printf("...\n");
